Homework-2/requirements.txt
flask
gunicorn
Flask-Failsafe
mysql-connector-python